nuevas ruta para usuario

+++ src/controllers/plantilla_insumoController.js
@@ -1,5 +1,6 @@
 const PlantillaInsumoService = require("../services/plantilla_insumoService");
 const dotenv = require("dotenv");
+const jwt = require("jsonwebtoken");
 
 dotenv.config();
 
@@ -20,6 +21,12 @@ const getAllPlantillasInsumos = async (req, res) => {
 };
 
 const getOnePlantillaInsumo = async (req, res) => {
+  if (!req.headers.authorization) {
+    return res.status(401).json({ error: "Not Authorized" });
+  }
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
   const {
     params: { planInId },
   } = req;
@@ -30,10 +37,12 @@ const getOnePlantillaInsumo = async (req, res) => {
     });
   }
   try {
-    const plantilla_insumo = await PlantillaInsumoService.getOnePlantillaInsumo(
-      planInId
-    );
-    res.send({ status: "OK", data: plantilla_insumo });
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user) {
+      const plantilla_insumo =
+        await PlantillaInsumoService.getOnePlantillaInsumo(planInId);
+      res.send({ status: "OK", data: plantilla_insumo });
+    }
   } catch (error) {
     res
       .status(error?.status || 500)
@@ -42,6 +51,12 @@ const getOnePlantillaInsumo = async (req, res) => {
 };
 
 const createNewPlantillaInsumo = async (req, res) => {
+  if (!req.headers.authorization) {
+    return res.status(401).json({ error: "Not Authorized" });
+  }
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
   const { body } = req;
   if (!body.nombre) {
     res.status(400).send({
@@ -59,9 +74,14 @@ const createNewPlantillaInsumo = async (req, res) => {
     image: body.image,
   };
   try {
-    const createdPlantillaInsumo =
-      await PlantillaInsumoService.createNewPlantillaInsumo(newPlantillaInsumo);
-    res.status(201).send({ status: "OK", data: createdPlantillaInsumo });
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user) {
+      const createdPlantillaInsumo =
+        await PlantillaInsumoService.createNewPlantillaInsumo(
+          newPlantillaInsumo
+        );
+      res.status(201).send({ status: "OK", data: createdPlantillaInsumo });
+    }
   } catch (error) {
     res
       .status(error?.status || 500)
@@ -70,6 +90,12 @@ const createNewPlantillaInsumo = async (req, res) => {
 };
 
 const updateOnePlantillaInsumo = async (req, res) => {
+  if (!req.headers.authorization) {
+    return res.status(401).json({ error: "Not Authorized" });
+  }
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
   const {
     body,
     params: { planInId },
@@ -81,9 +107,12 @@ const updateOnePlantillaInsumo = async (req, res) => {
     });
   }
   try {
-    const updatedPlantillaInsumo =
-      await PlantillaProductoService.updateOnePlantillaInsumo(planInId, body);
-    res.send({ status: "OK", data: updatedPlantillaInsumo });
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user) {
+      const updatedPlantillaInsumo =
+        await PlantillaProductoService.updateOnePlantillaInsumo(planInId, body);
+      res.send({ status: "OK", data: updatedPlantillaInsumo });
+    }
   } catch (error) {
     res
       .status(error?.status || 500)
@@ -92,13 +121,22 @@ const updateOnePlantillaInsumo = async (req, res) => {
 };
 
 const deleteOnePlantillaInsumo = async (req, res) => {
+  if (!req.headers.authorization) {
+    return res.status(401).json({ error: "Not Authorized" });
+  }
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
   const {
     params: { planInId },
   } = req;
   try {
-    const deletedPlantillaInsumo =
-      await PlantillaInsumoService.deleteOnePlantillaInsumo(planInId);
-    res.status(204).send({ status: "OK", data: deletedPlantillaInsumo });
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user) {
+      const deletedPlantillaInsumo =
+        await PlantillaInsumoService.deleteOnePlantillaInsumo(planInId);
+      res.status(204).send({ status: "OK", data: deletedPlantillaInsumo });
+    }
   } catch (error) {
     res
       .status(error?.status || 500)
