+++ src/controllers/plantilla_productoController.js
@@ -1,5 +1,6 @@
 const PlantillaProductoService = require("../services/plantilla_productoService");
 const dotenv = require("dotenv");
+const jwt = require("jsonwebtoken");
 
 dotenv.config();
 
@@ -20,6 +21,12 @@ const getAllPlantillasProductos = async (req, res) => {
 };
 
 const getOnePlantillaProducto = async (req, res) => {
+  if (!req.headers.authorization) {
+    return res.status(401).json({ error: "Not Authorized" });
+  }
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
   const {
     params: { planProId },
   } = req;
@@ -30,9 +37,12 @@ const getOnePlantillaProducto = async (req, res) => {
     });
   }
   try {
-    const plantilla_producto =
-      await PlantillaProductoService.getOnePlantillaProducto(planProId);
-    res.send({ status: "OK", data: plantilla_producto });
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user) {
+      const plantilla_producto =
+        await PlantillaProductoService.getOnePlantillaProducto(planProId);
+      res.send({ status: "OK", data: plantilla_producto });
+    }
   } catch (error) {
     res
       .status(error?.status || 500)
@@ -41,6 +51,12 @@ const getOnePlantillaProducto = async (req, res) => {
 };
 
 const createNewPlantillaProducto = async (req, res) => {
+  if (!req.headers.authorization) {
+    return res.status(401).json({ error: "Not Authorized" });
+  }
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
   const { body } = req;
   if (!body.nombre) {
     res.status(400).send({
@@ -58,11 +74,14 @@ const createNewPlantillaProducto = async (req, res) => {
     image: body.image,
   };
   try {
-    const createdPlantillaProducto =
-      await PlantillaProductoService.createNewPlantillaProducto(
-        newPlantillaProducto
-      );
-    res.status(201).send({ status: "OK", data: createdPlantillaProducto });
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user) {
+      const createdPlantillaProducto =
+        await PlantillaProductoService.createNewPlantillaProducto(
+          newPlantillaProducto
+        );
+      res.status(201).send({ status: "OK", data: createdPlantillaProducto });
+    }
   } catch (error) {
     res
       .status(error?.status || 500)
@@ -71,6 +90,12 @@ const createNewPlantillaProducto = async (req, res) => {
 };
 
 const updateOnePlantillaProducto = async (req, res) => {
+  if (!req.headers.authorization) {
+    return res.status(401).json({ error: "Not Authorized" });
+  }
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
   const {
     body,
     params: { planProId },
@@ -82,12 +107,15 @@ const updateOnePlantillaProducto = async (req, res) => {
     });
   }
   try {
-    const updatedPlantillaProducto =
-      await PlantillaProductoService.updateOnePlantillaProducto(
-        planProId,
-        body
-      );
-    res.send({ status: "OK", data: updatedPlantillaProducto });
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user) {
+      const updatedPlantillaProducto =
+        await PlantillaProductoService.updateOnePlantillaProducto(
+          planProId,
+          body
+        );
+      res.send({ status: "OK", data: updatedPlantillaProducto });
+    }
   } catch (error) {
     res
       .status(error?.status || 500)
@@ -96,13 +124,22 @@ const updateOnePlantillaProducto = async (req, res) => {
 };
 
 const deleteOnePlantillaProducto = async (req, res) => {
+  if (!req.headers.authorization) {
+    return res.status(401).json({ error: "Not Authorized" });
+  }
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
   const {
     params: { planProId },
   } = req;
   try {
-    const deletedPlantillaProducto =
-      await PlantillaProductoService.deleteOnePlantillaProducto(planProId);
-    res.status(204).send({ status: "OK", data: deletedPlantillaProducto });
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user) {
+      const deletedPlantillaProducto =
+        await PlantillaProductoService.deleteOnePlantillaProducto(planProId);
+      res.status(204).send({ status: "OK", data: deletedPlantillaProducto });
+    }
   } catch (error) {
     res
       .status(error?.status || 500)
